Move the hidraw/bpf backend handling into its own module
This is a bit messy because it's obviously moved over from file A to
file B and not yet properly restructured but at least it cleans up the
code a little bit.
