+++ src/main.rs
@@ -2,15 +2,10 @@
 
 use anyhow::{bail, Context, Result};
 use clap::{ColorChoice, Parser};
-use nix::poll::{poll, PollFd, PollFlags, PollTimeout};
 use owo_colors::{OwoColorize, Rgb, Stream::Stdout, Style};
-use std::cell::OnceCell;
 use std::collections::HashSet;
-use std::fs::OpenOptions;
-use std::io::{Read, Write};
+use std::io::Write;
 use std::ops::Deref;
-use std::os::fd::{AsFd, AsRawFd};
-use std::os::unix::fs::OpenOptionsExt;
 use std::path::{Path, PathBuf};
 use std::process::ExitCode;
 use std::sync::OnceLock;
@@ -25,34 +20,17 @@ use hidreport::hid::{
 };
 use hidreport::*;
 
-use libbpf_rs::libbpf_sys;
-use libbpf_rs::skel::OpenSkel as _;
-use libbpf_rs::skel::SkelBuilder as _;
-
-mod hidrecord {
-    include!(env!("SKELFILE"));
-}
-mod hidrecord_tracing {
-    include!(env!("SKELFILE_TRACING"));
-}
-
-use hidrecord::*;
-use hidrecord_tracing::*;
-
-mod hidrecording;
-mod libinput;
-
 static mut OUTFILE: OnceLock<
     std::sync::Mutex<std::cell::RefCell<std::io::LineWriter<std::fs::File>>>,
 > = OnceLock::new();
 
-enum Outfile {
+pub enum Outfile {
     Stdout,
     File(&'static mut std::sync::Mutex<std::cell::RefCell<std::io::LineWriter<std::fs::File>>>),
 }
 
 impl Outfile {
-    fn new() -> Self {
+    pub fn new() -> Self {
         unsafe {
             match OUTFILE.get_mut() {
                 None => Outfile::Stdout,
@@ -114,217 +92,6 @@ trait Backend {
     fn pid(&self) -> u32;
     fn rdesc(&self) -> &[u8];
     fn read_events(&self, use_bpf: BpfOption, rdesc: &ReportDescriptor) -> Result<()>;
-}
-
-struct HidrawBackend {
-    name: String,
-    bustype: u32,
-    vid: u32,
-    pid: u32,
-    rdesc: Vec<u8>,
-    device_path: Option<PathBuf>,
-}
-
-impl HidrawBackend {
-    fn read_events_loop(
-        &self,
-        path: &Path,
-        rdesc: &ReportDescriptor,
-        map_ringbuf: Option<&libbpf_rs::Map>,
-    ) -> Result<()> {
-        let mut f = OpenOptions::new()
-            .read(true)
-            .custom_flags(libc::O_NONBLOCK)
-            .open(path)?;
-
-        let timeout = PollTimeout::try_from(1000).unwrap();
-        let start_time: OnceCell<Instant> = OnceCell::new();
-        let mut last_timestamp: Option<Instant> = None;
-        let mut data = [0; 1024];
-        let mut bpf_vec = Vec::new();
-
-        let ringbuf = map_ringbuf.map(|map_ringbuf| {
-            let mut builder = libbpf_rs::RingBufferBuilder::new();
-            builder
-                .add(map_ringbuf, |data| {
-                    bpf_event_handler(data, &mut bpf_vec, &start_time)
-                })
-                .unwrap();
-            builder.build().unwrap()
-        });
-
-        loop {
-            let mut pollfds = vec![PollFd::new(f.as_fd(), PollFlags::POLLIN)];
-            if let Some(ref ringbuf) = ringbuf {
-                let ringbuf_fd = unsafe {
-                    std::os::fd::BorrowedFd::borrow_raw(ringbuf.epoll_fd() as std::os::fd::RawFd)
-                };
-                pollfds.push(PollFd::new(ringbuf_fd, PollFlags::POLLIN));
-            }
-
-            if poll(&mut pollfds, timeout)? > 0 {
-                let has_events: Vec<bool> = pollfds
-                    .iter()
-                    .map(|fd| fd.revents())
-                    .map(|revents| revents.map_or(false, |flag| flag.intersects(PollFlags::POLLIN)))
-                    .collect();
-
-                if has_events[0] {
-                    match f.read(&mut data) {
-                        Ok(_nbytes) => {
-                            last_timestamp = print_current_time(last_timestamp);
-                            let _ = start_time.get_or_init(|| last_timestamp.unwrap());
-                            print_input_report_description(&data, rdesc)?;
-
-                            let elapsed = start_time.get().unwrap().elapsed();
-                            // This prints the B: 123 00 01 02 ... data line via the callback
-                            if let Some(ref ringbuf) = ringbuf {
-                                let _ = ringbuf.consume();
-                            }
-
-                            print_input_report_data(&data, rdesc, &elapsed)?;
-                        }
-                        Err(e) => {
-                            if e.kind() != std::io::ErrorKind::WouldBlock {
-                                bail!(e);
-                            }
-                        }
-                    };
-                }
-                if *has_events.get(1).unwrap_or(&false) {
-                    if let Some(ref ringbuf) = ringbuf {
-                        last_timestamp = print_current_time(last_timestamp);
-                        let _ = start_time.get_or_init(|| last_timestamp.unwrap());
-                        let _ = ringbuf.consume();
-                    }
-                }
-            } else if last_timestamp.is_some() {
-                print_current_time(last_timestamp);
-            }
-        }
-    }
-}
-
-impl TryFrom<&Path> for HidrawBackend {
-    type Error = anyhow::Error;
-
-    fn try_from(path: &Path) -> Result<Self> {
-        if ["/dev", "/sys"]
-            .iter()
-            .any(|prefix| path.starts_with(prefix))
-        {
-            let sysfs = find_sysfs_path(path)?;
-            let rdesc_path = sysfs.join("report_descriptor");
-            if !rdesc_path.exists() {
-                bail!("Unable to find report descriptor at {rdesc_path:?}");
-            }
-
-            let (name, ids) = parse_uevent(&sysfs)?;
-            let (bustype, vid, pid) = ids;
-
-            let bytes = std::fs::read(&rdesc_path)?;
-            if bytes.is_empty() {
-                bail!("Empty report descriptor");
-            }
-
-            let device_path = if path.starts_with("/dev") {
-                Some(PathBuf::from(path))
-            } else {
-                None
-            };
-
-            Ok(HidrawBackend {
-                name,
-                bustype,
-                vid,
-                pid,
-                rdesc: bytes,
-                device_path,
-            })
-        } else {
-            bail!("Not a syfs file or hidraw node");
-        }
-    }
-}
-
-impl Backend for HidrawBackend {
-    fn name(&self) -> &str {
-        &self.name
-    }
-
-    fn bustype(&self) -> u32 {
-        self.bustype
-    }
-
-    fn vid(&self) -> u32 {
-        self.vid
-    }
-
-    fn pid(&self) -> u32 {
-        self.pid
-    }
-
-    fn rdesc(&self) -> &[u8] {
-        &self.rdesc
-    }
-
-    fn read_events(&self, use_bpf: BpfOption, rdesc: &ReportDescriptor) -> Result<()> {
-        if self.device_path.is_none() {
-            return Ok(());
-        }
-        let path = self.device_path.as_ref().unwrap();
-        match preload_bpf_tracer(use_bpf, &path)? {
-            HidBpfSkel::None => self.read_events_loop(&path, rdesc, None)?,
-            HidBpfSkel::StructOps(skel) => {
-                let maps = skel.maps();
-                // We need to keep _link around or the program gets immediately removed
-                let _link = maps.hid_record().attach_struct_ops()?;
-                self.read_events_loop(&path, rdesc, Some(maps.events()))?
-            }
-            HidBpfSkel::Tracing(skel, hid_id) => {
-                let attach_args = attach_prog_args {
-                    prog_fd: skel.progs().hid_record_event().as_fd().as_raw_fd(),
-                    hid: hid_id,
-                    retval: -1,
-                };
-
-                let _link =
-                    run_syscall_prog_attach(skel.progs().attach_prog(), attach_args).unwrap();
-                self.read_events_loop(&path, rdesc, Some(skel.maps().events()))?
-            }
-        }
-        Ok(())
-    }
-}
-
-const PACKET_SIZE: usize = 64;
-
-// A Rust version of hid_recorder_event in hidrecord.bpf.c
-// struct hid_recorder_event {
-// 	__u8 length;
-// 	__u8 data[64];
-// };
-#[allow(non_camel_case_types)]
-#[repr(C)]
-struct hid_recorder_event {
-    packet_count: u8,
-    packet_number: u8,
-    length: u8,
-    data: [u8; PACKET_SIZE],
-}
-
-// A Rust version of attach_prog_args in hidrecord_tracing.bpf.c
-// struct attach_prog_args {
-// 	int prog_fd;
-// 	unsigned int hid;
-// 	int retval;
-// };
-#[allow(non_camel_case_types)]
-#[repr(C)]
-struct attach_prog_args {
-    prog_fd: i32,
-    hid: u32,
-    retval: i32,
 }
 
 #[derive(Default)]
@@ -395,6 +162,10 @@ macro_rules! cprint {
     }};
 }
 
+mod hidraw;
+mod hidrecording;
+mod libinput;
+
 #[derive(Parser, Debug)]
 #[command(version, about, long_about = None)]
 struct Cli {
@@ -432,31 +203,6 @@ struct Options {
     full: bool,
     only_describe: bool,
     bpf: ColorChoice,
-}
-
-#[derive(Debug)]
-pub enum BpfError {
-    LibBPFError { error: libbpf_rs::Error },
-    OsError { errno: u32 },
-}
-
-impl std::error::Error for BpfError {}
-
-impl std::fmt::Display for BpfError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            BpfError::LibBPFError { error } => write!(f, "{error}"),
-            BpfError::OsError { errno } => {
-                write!(f, "{}", libbpf_rs::Error::from_raw_os_error(*errno as i32))
-            }
-        }
-    }
-}
-
-impl From<libbpf_rs::Error> for BpfError {
-    fn from(e: libbpf_rs::Error) -> BpfError {
-        BpfError::LibBPFError { error: e }
-    }
 }
 
 fn fmt_main_item(item: &MainItem) -> String {
@@ -672,7 +418,7 @@ fn print_rdesc_items(bytes: &[u8]) -> Result<()> {
 }
 
 // This would be easier with udev but let's keep the dependencies relatively minimal.
-fn find_sysfs_path(path: &Path) -> Result<PathBuf> {
+pub fn find_sysfs_path(path: &Path) -> Result<PathBuf> {
     let pathstr = path.to_string_lossy();
     let sysfs: PathBuf;
     if pathstr.starts_with("/dev/hidraw") {
@@ -1049,7 +795,7 @@ fn print_report_summary(r: &impl Report, opts: &Options) {
     }
 }
 
-fn parse_uevent(sysfs: &Path) -> Result<(String, (u32, u32, u32))> {
+pub fn parse_uevent(sysfs: &Path) -> Result<(String, (u32, u32, u32))> {
     // uevent should contain
     // HID_NAME=foo bar
     // HID_ID=00003:0002135:0000123513
@@ -1276,7 +1022,7 @@ pub fn print_input_report_description(bytes: &[u8], rdesc: &ReportDescriptor) ->
     Ok(())
 }
 
-fn print_input_report_data(
+pub fn print_input_report_data(
     bytes: &[u8],
     rdesc: &ReportDescriptor,
     elapsed: &Duration,
@@ -1299,7 +1045,7 @@ fn print_input_report_data(
     Ok(())
 }
 
-fn print_bpf_input_report_data(bytes: &[u8], elapsed: &Duration) {
+pub fn print_bpf_input_report_data(bytes: &[u8], elapsed: &Duration) {
     let bytes = bytes
         .iter()
         .fold("".to_string(), |acc, b| format!("{acc}{b:02x} "));
@@ -1313,7 +1059,7 @@ fn print_bpf_input_report_data(bytes: &[u8], elapsed: &Duration) {
     );
 }
 
-fn print_current_time(last_timestamp: Option<Instant>) -> Option<Instant> {
+pub fn print_current_time(last_timestamp: Option<Instant>) -> Option<Instant> {
     let prev_timestamp = last_timestamp.unwrap_or(Instant::now());
     let elapsed = prev_timestamp.elapsed().as_secs();
     let now = chrono::prelude::Local::now();
@@ -1359,152 +1105,6 @@ fn find_device() -> Result<PathBuf> {
     Ok(path)
 }
 
-fn bpf_event_handler(
-    data: &[u8],
-    buffer: &mut Vec<u8>,
-    start_time: &OnceCell<Instant>,
-) -> ::std::os::raw::c_int {
-    if data.len() != std::mem::size_of::<hid_recorder_event>() {
-        eprintln!(
-            "Invalid size {} != {}",
-            data.len(),
-            std::mem::size_of::<hid_recorder_event>()
-        );
-        return 1;
-    }
-
-    let event = unsafe { &*(data.as_ptr() as *const hid_recorder_event) };
-
-    if event.length == 0 {
-        return 1;
-    }
-
-    let elapsed = start_time.get().unwrap().elapsed();
-
-    let size = if event.packet_number == event.packet_count - 1 {
-        event.length as usize - event.packet_number as usize * PACKET_SIZE
-    } else {
-        PACKET_SIZE
-    };
-
-    if event.packet_number == 0 {
-        buffer.clear();
-    }
-
-    buffer.extend_from_slice(&event.data[..size]);
-
-    if event.packet_number == event.packet_count - 1 {
-        print_bpf_input_report_data(&buffer, &elapsed);
-    }
-    0
-}
-
-enum HidBpfSkel {
-    None,
-    StructOps(Box<HidrecordSkel<'static>>),
-    Tracing(Box<HidrecordTracingSkel<'static>>, u32),
-}
-
-fn print_to_log(level: libbpf_rs::PrintLevel, msg: String) {
-    /* we strip out the 3 following lines that happen when the kernel
-     * doesn't support HID-BPF struct_ops
-     */
-    let ignore_msgs = [
-        "struct bpf_struct_ops_hid_bpf_ops is not found in kernel BTF",
-        "failed to load object 'hidrecord_bpf'",
-        "failed to load BPF skeleton 'hidrecord_bpf': -2",
-    ];
-    if ignore_msgs.iter().any(|ignore| msg.contains(ignore)) {
-        return;
-    }
-    match level {
-        libbpf_rs::PrintLevel::Info => cprintln!(Styles::Bpf, "{}", msg.trim()),
-        libbpf_rs::PrintLevel::Warn => cprintln!(Styles::Note, "{}", msg.trim()),
-        _ => (),
-    }
-}
-
-fn preload_bpf_tracer(use_bpf: BpfOption, path: &Path) -> Result<HidBpfSkel> {
-    let sysfs = find_sysfs_path(path)?.canonicalize()?;
-    let hid_id = u32::from_str_radix(
-        sysfs
-            .extension()
-            .unwrap()
-            .to_str()
-            .expect("not a hex value"),
-        16,
-    )
-    .unwrap();
-
-    let sysfs_name = sysfs
-        .file_name()
-        .unwrap()
-        .to_str()
-        .unwrap()
-        .replace([':', '.'], "_");
-
-    let bpffs = PathBuf::from("/sys/fs/bpf/hid/").join(sysfs_name);
-
-    let enable_bpf = match use_bpf {
-        BpfOption::Never => false,
-        BpfOption::Always => true,
-        BpfOption::Auto => bpffs.exists(),
-    };
-
-    if !enable_bpf {
-        return Ok(HidBpfSkel::None);
-    }
-
-    libbpf_rs::set_print(Some((libbpf_rs::PrintLevel::Info, print_to_log)));
-
-    let skel_builder = HidrecordSkelBuilder::default();
-    let mut open_skel = skel_builder.open().unwrap();
-    let hid_record_update = open_skel.struct_ops.hid_record_mut();
-    hid_record_update.hid_id = hid_id as i32;
-
-    if let Ok(skel) = open_skel.load() {
-        return Ok(HidBpfSkel::StructOps(Box::new(skel)));
-    }
-
-    let skel_builder = HidrecordTracingSkelBuilder::default();
-    let skel = skel_builder.open()?.load()?;
-    Ok(HidBpfSkel::Tracing(Box::new(skel), hid_id))
-}
-
-fn run_syscall_prog_generic<T>(prog: &libbpf_rs::Program, data: T) -> Result<T, BpfError> {
-    let fd = prog.as_fd().as_raw_fd();
-    let data_ptr: *const libc::c_void = &data as *const _ as *const libc::c_void;
-    let mut run_opts = libbpf_sys::bpf_test_run_opts {
-        sz: std::mem::size_of::<libbpf_sys::bpf_test_run_opts>()
-            .try_into()
-            .unwrap(),
-        ctx_in: data_ptr,
-        ctx_size_in: std::mem::size_of::<T>() as u32,
-        ..Default::default()
-    };
-
-    let run_opts_ptr: *mut libbpf_sys::bpf_test_run_opts = &mut run_opts;
-
-    match unsafe { libbpf_sys::bpf_prog_test_run_opts(fd, run_opts_ptr) } {
-        0 => Ok(data),
-        e => Err(BpfError::OsError { errno: -e as u32 }),
-    }
-}
-
-fn run_syscall_prog_attach(
-    prog: &libbpf_rs::Program,
-    attach_args: attach_prog_args,
-) -> Result<i32, BpfError> {
-    let args = run_syscall_prog_generic(prog, attach_args)?;
-    if args.retval < 0 {
-        Err(BpfError::OsError {
-            errno: -args.retval as u32,
-        })
-    } else {
-        Ok(args.retval)
-    }
-}
-
 fn process(backend: impl Backend, opts: &Options) -> Result<()> {
     let rdesc = parse_report_descriptor(&backend, &opts)?;
     if !opts.only_describe {
@@ -1538,7 +1138,7 @@ fn hid_recorder() -> Result<()> {
         only_describe: cli.only_describe,
         bpf: cli.bpf,
     };
-    if let Ok(backend) = HidrawBackend::try_from(path.as_path()) {
+    if let Ok(backend) = hidraw::HidrawBackend::try_from(path.as_path()) {
         process(backend, &opts)
     } else if let Ok(backend) = libinput::LibinputRecordingBackend::try_from(path.as_path()) {
         process(backend, &opts)
@@ -1573,7 +1173,7 @@ mod tests {
             .filter(|name| name.starts_with("hidraw"))
             .collect();
         for hidraw in hidraws.iter().map(|h| PathBuf::from("/dev/").join(h)) {
-            let result = HidrawBackend::try_from(hidraw.as_path());
+            let result = hidraw::HidrawBackend::try_from(hidraw.as_path());
             assert!(result.is_ok());
         }
     }
@@ -1595,7 +1195,7 @@ mod tests {
             assert!(evdevs
                 .iter()
                 .map(|n| PathBuf::from("/dev/input").join(n))
-                .any(|evdev| HidrawBackend::try_from(evdev.as_path()).is_ok()));
+                .any(|evdev| hidraw::HidrawBackend::try_from(evdev.as_path()).is_ok()));
         }
     }
 
@@ -1615,7 +1215,7 @@ mod tests {
             .map(|path| {
                 (
                     path.clone(),
-                    HidrawBackend::try_from(path.as_path()).unwrap(),
+                    hidraw::HidrawBackend::try_from(path.as_path()).unwrap(),
                 )
             })
         {
